Assorted world/board save fixes.
* Added backup system. Now, when a world is being saved, it is first renamed
  (to f.e. UNTITLED.ZZT.BAK), then saved, then the backup is removed only if
  the saving process was successful. This means that in the event of edge
  cases causing the save to fail (see below) or KevEdit crashes (as Dr_Dos
  and myself have recently encountered but not reproduced), the world data
  should be safe. Note that I did not make it configurable, at this time.
* Fixed board size information for boards over 65535 bytes. (For example, try
  placing an object with a fair amount of text, and filling the board with
  it - it can easily go into the hundreds of kilobytes.)
* Saving boards and worlds above 65535 bytes is now strictly forbidden and
  will cause an error - it is not possible to save such a world uncorrupted
  in the ZZT world format, due to the 16-bit board size limit.
* Errors are now caught and the user is notified about them during board
  and world saving.
* Minor cleanups around str_dup's documentation and usage.

+++ src/libzzt2/zzt.h
@@ -325,7 +325,7 @@ int zztBoardCompress(ZZTboard *board);
 /* zztBoardGetSize(board)
  * Determine the size of a board in bytes
  */
-uint16_t zztBoardGetSize(ZZTboard *board);
+uint32_t zztBoardGetSize(ZZTboard *board);
 /* zztWorldAddBoard(world, title)
  * Create a blank board at the end of the world with the given title
  */
